backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
azure-cognitiveservices-speech
python-dotenv
